Update with new tickets counts
New ticket counts beginning with 2015 race lottery:

Each runner who enters the lottery and fails to gain entry into the Run (and
otherwise doesn’t gain an entry via other means such as an aid station,
sponsor, or MUC spot) will have additional tickets in the hat when entering
the lottery the following year, thus improving the probability of being
selected. Every lottery applicant will receive 2^(n-1) tickets in the hat for
each consecutive failure to gain entry where n is the number of years applied.
As an example: If a runner applies and fails to gain entry two consecutive
times, they will have four tickets in the hat when entered in the lottery the
third consecutive year (2^2), eight in fourth consecutive year (2^3), 16 in
fifth (2^4), 32 in sixth (2^5), etc.  Conversely, if a runner enters the
lottery one year, is not chosen and does not enter the following year, they
will not have a extra tickets in the hat when applying the year after that.